MSUtil renamed to MSUtility
