+++ MobileCenter/MobileCenter/Internals/Channel/MSLogManagerDefault.m
@@ -6,6 +6,7 @@
 #import "MSLogManagerDefaultPrivate.h"
 #import "MSMobileCenterErrors.h"
 #import "MobileCenter+Internal.h"
+#import "MSUtility+Date.h"
 
 static char *const MSlogsDispatchQueue = "com.microsoft.azure.mobile.mobilecenter.LogManagerQueue";
 
@@ -106,7 +107,7 @@ static char *const MSlogsDispatchQueue = "com.microsoft.azure.mobile.mobilecente
   id<MSChannel> channel = [self channelForPriority:priority];
 
   // Set common log info.
-  log.toffset = [NSNumber numberWithLongLong:[MSUtil nowInMilliseconds]];
+  log.toffset = [NSNumber numberWithLongLong:[MSUtility nowInMilliseconds]];
 
   // Only add device info in case the log doesn't have one. In case the log is restored after a crash or for crashes,
   // We don't want the device information to be updated but want the old one preserved.
