--- MobileCenter/MobileCenter/Internals/Utils/MSUtilPrivate.h
@@ -1,17 +1,0 @@
-#import "MSUtil.h"
-
-#import <UIKit/UIKit.h>
-
-@interface MSUtil ()
-
-/**
- * Get the shared app state.
- *
- * @discussion This method is exposed for testing purposes. The shared app state is resolved at runtime by this method
- * which makes the UIApplication not mockable. This method is meant to be stubbed in tests to inject the desired
- * application states.
- * @return The shared app state.
- */
-+ (UIApplicationState)sharedAppState;
-
-@end
