+++ MobileCenter/MobileCenter/Internals/Utils/MSUtility+Application.m
@@ -10,4 +10,12 @@
     return MSApplicationStateUnknown;
 }
 
++ (UIApplicationState)sharedAppState {
+    
+    // Compute selector at runtime for more discretion.
+    SEL sharedAppSel = NSSelectorFromString(@"sharedApplication");
+    return [[((UIApplication *(*)(id, SEL)) [[UIApplication class] methodForSelector:sharedAppSel])(
+                                                                                                    [UIApplication class], sharedAppSel) valueForKey:@"applicationState"] longValue];
+}
+
 @end
