+++ MobileCenter/MobileCenter/Internals/Utils/MSUtility.m
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ */
+
+#import "MSUtility.h"
+
+@implementation MSUtility
+@end
