+++ MobileCenter/MobileCenterTests/MSIngestionSenderTests.m
@@ -12,6 +12,7 @@
 #import "MSSenderCall.h"
 #import "MSSenderDelegate.h"
 #import "MobileCenter+Internal.h"
+#import "MSUtility+Date.h"
 
 #import "OCMock.h"
 #import <OCHamcrestIOS/OCHamcrestIOS.h>
@@ -367,7 +368,7 @@ static NSString *const kMSAppSecret = @"mockAppSecret";
 
   MSMockLog *log1 = [[MSMockLog alloc] init];
   log1.sid = MS_UUID_STRING;
-  log1.toffset = [NSNumber numberWithLongLong:[MSUtil nowInMilliseconds]];
+  log1.toffset = [NSNumber numberWithLongLong:[MSUtility nowInMilliseconds]];
 
   // Log does not have device info, therefore, it's an invalid log
   MSLogContainer *container = [[MSLogContainer alloc] initWithBatchId:@"1" andLogs:(NSArray<MSLog> *)@[ log1 ]];
@@ -578,12 +579,12 @@ static NSString *const kMSAppSecret = @"mockAppSecret";
 
   MSMockLog *log1 = [[MSMockLog alloc] init];
   log1.sid = MS_UUID_STRING;
-  log1.toffset = [NSNumber numberWithLongLong:[MSUtil nowInMilliseconds]];
+  log1.toffset = [NSNumber numberWithLongLong:[MSUtility nowInMilliseconds]];
   log1.device = device;
 
   MSMockLog *log2 = [[MSMockLog alloc] init];
   log2.sid = MS_UUID_STRING;
-  log2.toffset = [NSNumber numberWithLongLong:[MSUtil nowInMilliseconds]];
+  log2.toffset = [NSNumber numberWithLongLong:[MSUtility nowInMilliseconds]];
   log2.device = device;
 
   MSLogContainer *logContainer =
