+++ MobileCenterAnalytics/MobileCenterAnalyticsTests/MSEventLogTests.m
@@ -5,6 +5,7 @@
 
 #import "MSEventLog.h"
 #import "MobileCenter+Internal.h"
+#import "MSUtility+Date.h"
 
 @interface MSEventLogTests : XCTestCase
 
@@ -38,7 +39,7 @@
   MSDevice *device = [MSDevice new];
   NSString *sessionId = @"1234567890";
   NSDictionary *properties = @{ @"Key" : @"Value" };
-  long long createTime = [MSUtil nowInMilliseconds];
+  long long createTime = [MSUtility nowInMilliseconds];
   NSNumber *tOffset = @(createTime);
 
   self.sut.eventId = eventId;
@@ -61,7 +62,7 @@
   assertThat(actual[@"properties"], equalTo(properties));
   assertThat(actual[@"device"], notNilValue());
   NSTimeInterval seralizedToffset = [actual[@"toffset"] longLongValue];
-  NSTimeInterval actualToffset = [MSUtil nowInMilliseconds] - createTime;
+  NSTimeInterval actualToffset = [MSUtility nowInMilliseconds] - createTime;
   assertThat(@(seralizedToffset), lessThan(@(actualToffset)));
 }
