+++ MobileCenterAnalytics/MobileCenterAnalyticsTests/MSPageLogTests.m
@@ -4,6 +4,7 @@
 #import <XCTest/XCTest.h>
 
 #import "MSPageLog.h"
+#import "MSUtility+Date.h"
 
 @interface MSPageLogTests : XCTestCase
 
@@ -36,7 +37,7 @@
   MSDevice *device = [MSDevice new];
   NSString *sessionId = @"1234567890";
   NSDictionary *properties = @{ @"Key" : @"Value" };
-  long long createTime = [MSUtil nowInMilliseconds];
+  long long createTime = [MSUtility nowInMilliseconds];
   NSNumber *tOffset = [NSNumber numberWithLongLong:createTime];
 
   self.sut.name = pageName;
@@ -57,7 +58,7 @@
   assertThat(actual[@"properties"], equalTo(properties));
   assertThat(actual[@"device"], notNilValue());
   NSTimeInterval seralizedToffset = [actual[@"toffset"] longLongValue];
-  NSTimeInterval actualToffset = [MSUtil nowInMilliseconds] - createTime;
+  NSTimeInterval actualToffset = [MSUtility nowInMilliseconds] - createTime;
   assertThat(@(seralizedToffset), lessThan(@(actualToffset)));
 
 }
